backend/pyproject.toml
[project]
name = "clouisle-backend"
version = "0.1.0"
description = "Backend for Clouisle application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "tortoise-orm[asyncpg]>=0.20.0",
    "asyncpg>=0.29.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "email-validator>=2.3.0",
    "bcrypt==4.0.1",
    "aiosmtplib>=3.0.0",
    "aiofiles>=25.1.0",
    "types-aiofiles>=25.1.0.20251011",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-anthropic>=1.3.0",
    "langgraph>=1.0.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
